Add Wasm Local component + locals_map on Funcs
The Local component replaces both the Local instruction (which isn't actually an instruction) and the FuncParam Component (which had the exact same signature as a Local).

The new locals_map stores all the locals in a function and hashes them by type. This makes Wasm binary generation a lot easier.