Update CFGFactsEquality to be more strict
This update takes advantage of the new BB IL test case format and allows CFGFactsEquality to check for identical block IDs between the actual and expected CFGFacts. This removes the ambiguity that could arise from CFG nodes being ordered differently in the internal list (which does not actually effect correctness).